Allow layout selection for basic Markdown content
I opted to only allow layout selection in basic Markdown content because that
is the only place I have need of it. Once the need arises, I will cary the
implementation forward, even though it could be potentially confusing that
Markdown files can select a layout but haml or other files cannot.